relocating necessary jackrabbit bits to a new brix-jackrabbit module; BrixUrlCodingStrategy is last in brix-core to rely on jackrabbit (because of a exception); rest of jackrabbit in other modules as DEMO are nearly all test-only! (1 jackrabbit-core in brix-wrappers still there because brix-core needs for BrixUrlCodingStrategy)
