+++ brix-demo/pom.xml
@@ -147,6 +147,10 @@
    <artifactId>brix-workspace</artifactId>
   </dependency>
   <dependency>
+      <groupId>brix</groupId>
+      <artifactId>brix-jackrabbit</artifactId>
+  </dependency>
+  <dependency>
    <groupId>org.apache.wicket</groupId>
    <artifactId>wicket</artifactId>
   </dependency>
