+++ brix-jackrabbit/src/main/java/brix/util/RmiRepositoryFactory.java
