+++ brix-wrapper/pom.xml
@@ -38,6 +38,7 @@
      <artifactId>log4j</artifactId>
     </exclusion>
    </exclusions>
+   <scope>test</scope>
   </dependency>
   <dependency>
    <groupId>org.apache.jackrabbit</groupId>
@@ -52,6 +53,7 @@
      <artifactId>log4j</artifactId>
     </exclusion>
    </exclusions>
+
   </dependency>
   <dependency>
    <groupId>org.apache.jackrabbit</groupId>
@@ -66,6 +68,7 @@
      <artifactId>log4j</artifactId>
     </exclusion>
    </exclusions>
+      <scope>test</scope>
   </dependency>
   <dependency>
    <groupId>org.apache.jackrabbit</groupId>
@@ -80,6 +83,7 @@
      <artifactId>log4j</artifactId>
     </exclusion>
    </exclusions>
+      <scope>test</scope>
   </dependency>
  </dependencies>
  <build>
